requirements.txt
python-dotenv
groq
PyMuPDF
requests
beautifulsoup4